Add support for timezone conversion
This changeset is rather large because it encompasses some
functionality that was required to get to this state.

* TimeZone::from_utc was added to convert a UTC DateTime
* DateTime::shift was added for UTC DateTimes to add from an offset.
* A bunch of tests (adapted from the Python repo) were added

The TimeZone::from_utc method *could* potentially be in a separate
trait named FromUtc. However, the ergonomics of defining two separate
traits seemed negative so I opted to keep it within the same trait.

The DateTime::shift method is unfortunately there for necessity of
implementation in TimeZone::from_utc. The naming is also unfortunate
since it sounds more general than it actually is.

The DateTime::to_utc method is currently crate private though it might
be useful to export it elsewhere.