Add support for retrieving the local time and timezone
This unfortunately uses the "infamous" localtime_r API that made waves
last year as the source of CVEs. There is no replacement API for this
under POSIX so we're at a standstill.

After reading numerous issues, forum posts, and reddit posts on the
topic I eventually came to the conclusion that `localtime_r` and the
other related POSIX APIs weren't exactly the problem, but `setenv` was
instead. The `setenv` call being not thread-safe essentially makes it
a foot-gun for any API that uses `getenv` regardless of its status as
a re-entrant API.

A rewrite to pure Rust would solve the issue since the standard
library's `getenv` API makes a copy of the buffer and is thus free
from the invalidation that occurs when a `setenv` is called. However,
rewriting these APIs in Rust is a large undertaking.

I believe the only solution forward, at the moment, is to mark the
`std::env::set_var` API as `unsafe` and it seems that the standard
library team is converging to that decision as well. When that time
comes, any chance of doing a segfault using purely safe Rust is no
longer possible and these APIs will hopefully lose their infamy. At
least, for this particular issue.