Simplify TimeZone trait
The original trait was modeled after the Python `tzinfo` class, to a
fault. The translation was unfortunately complicated and cumbersome
to use. After dabbling with integration with the TZDB I realised that
certain changes had to be done to be done:

* The bounds on Tz propagated on every method even though it was
  not necessary.
* The dst_offset method was essentially meaningless since the offset
  method contained the same information.

A consequence of this redesign is that users will need to override the
datetime_at method to actually convert from Utc to the current
timezone. In the future there will be more methods added into the
TimeZone trait dealing with ambiguous and imaginary times and more
conversions which will probably complicate the TimeZone trait even
further.

There is probably no good way to do this since timezones are inherently
tricky.