Split format feature into two separate features
The two features are now "parsing" and "formatting".

This also changes the dependence of the format_spec proc macro into the
macros feature, which admittedly makes more sense though it is
unfortunate that there's no way to only enable it as a dependency if
`("formatting" || "parsing") && "macros"` is enabled. This seems like
a Cargo limitation for now.