Refactor timezone transition data layer
The old code felt a bit more complicated than I wanted it to be. This
new data layer is based off of the fact that time generally goes in a
straight line if looking at it from an instant point of view. Since
the data is already sorted this format makes immediate sense without
having to muck with having a previous element to keep track of.