Store UtcOffset as a member inside DateTime
This should save some potential computation for more relatively
expensive UtcOffset lookups (e.g. TZDB). This comes at a price of
some extra memory per DateTime instance but the simplicity of the code
and not having to call the offset calculation function every time a
comparison is done or some arithmetic makes up for this fact.