Add TimeZone::resolve to resolve a local time
This allows users of the library to handle whether a time is ambiguous
or missing. In the future this will be a building block for other
methods, such as `at` or `at_strictly`. Right now this isn't
implemented in the `eos_tz::TimeZone` type since that type has some
surprising flaws that require refactoring, however this is a step in
the right direction.