Change TimeZone::name and TimeZone::offset to work with ambiguity
The prior definitions using a (date, time) pair made it prone for
misuse when faced with an ambiguous or missing time. The new trait
method definition uses the new Timestamp trait to point to an "instant"
of time and allows for lookups to be unambiguous.

In order to test this, everything now tests whether the timezone names
are as expected. So far all these tests pass, making it so timezone
support should be complete.