Use an inner Arc for timezone data
This makes the type still cloneable but makes the clone cheaper to do
since it'd just be incrementing the inner Arc's reference count. The
library does indiscriminate copying of timezones quite often since the
datetime type is immutable so having a cheap to clone type is necessary
for performance reasons.