Reduce Interval size significantly
This reduces the size of Interval from 38 bytes to 16 bytes at the cost
of some precision (nanoseconds -> microseconds). This also makes it so
the units are always retrieved in normalized form. Due to complications
of the new inner format, this removes certain modifiers like `with_`
methods and the now redundant `normalize()` method.